Removed impl from BucketManager
