+++ src/bucket/BucketManager.h
@@ -1,161 +1,42 @@
 #pragma once
 
-// Copyright 2014 Stellar Development Foundation and contributors. Licensed
+#include "bucket/Bucket.h"
+#include "bucket/BucketList.h"
+#include "bucket/BucketMergeMap.h"
+#include "util/types.h"
+#include "xdr/Stellar-ledger.h"
+
+#include <map>
+#include <memory>
+#include <mutex>
+#include <set>
+#include <string>
+
+// Copyright 2015 Stellar Development Foundation and contributors. Licensed
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "bucket/Bucket.h"
-#include "util/NonCopyable.h"
-#include "util/types.h"
-#include <future>
-#include <map>
-#include <memory>
-#include <optional>
-#include <set>
-
-#include "medida/timer_context.h"
-
 namespace medida
 {
+class Timer;
 class Meter;
+class Counter;
 }
 
 namespace stellar
 {
 
+class TmpDir;
 class AbstractLedgerTxn;
 class Application;
-class BasicWork;
+class Bucket;
 class LiveBucketList;
-class HotArchiveBucketList;
 class BucketSnapshotManager;
-class Config;
 class SearchableLiveBucketListSnapshot;
-class TmpDirManager;
+struct BucketEntryCounters;
+enum class LedgerEntryTypeAndDurability : uint32_t;
+
 struct HistoryArchiveState;
-struct InflationWinner;
-struct LedgerHeader;
-struct MergeKey;
-
-// A fine-grained merge-operation-counter structure for tracking various
-// events during merges. These are not medida counters because we do not
-// want or need to publish this level of granularity outside of testing, and
-// we do want merges to run as quickly as possible.
-struct MergeCounters
-{
-    uint64_t mPreInitEntryProtocolMerges{0};
-    uint64_t mPostInitEntryProtocolMerges{0};
-
-    uint64_t mRunningMergeReattachments{0};
-    uint64_t mFinishedMergeReattachments{0};
-
-    uint64_t mPreShadowRemovalProtocolMerges{0};
-    uint64_t mPostShadowRemovalProtocolMerges{0};
-
-    uint64_t mNewMetaEntries{0};
-    uint64_t mNewInitEntries{0};
-    uint64_t mNewLiveEntries{0};
-    uint64_t mNewDeadEntries{0};
-    uint64_t mOldMetaEntries{0};
-    uint64_t mOldInitEntries{0};
-    uint64_t mOldLiveEntries{0};
-    uint64_t mOldDeadEntries{0};
-
-    uint64_t mOldEntriesDefaultAccepted{0};
-    uint64_t mNewEntriesDefaultAccepted{0};
-    uint64_t mNewInitEntriesMergedWithOldDead{0};
-    uint64_t mOldInitEntriesMergedWithNewLive{0};
-    uint64_t mOldInitEntriesMergedWithNewDead{0};
-    uint64_t mNewEntriesMergedWithOldNeitherInit{0};
-
-    uint64_t mShadowScanSteps{0};
-    uint64_t mMetaEntryShadowElisions{0};
-    uint64_t mLiveEntryShadowElisions{0};
-    uint64_t mInitEntryShadowElisions{0};
-    uint64_t mDeadEntryShadowElisions{0};
-
-    uint64_t mOutputIteratorTombstoneElisions{0};
-    uint64_t mOutputIteratorBufferUpdates{0};
-    uint64_t mOutputIteratorActualWrites{0};
-    MergeCounters& operator+=(MergeCounters const& delta);
-    bool operator==(MergeCounters const& other) const;
-};
-
-// Stores key that is eligible for eviction and the position of the eviction
-// iterator as if that key was the last entry evicted
-struct EvictionResultEntry
-{
-    LedgerKey key;
-    EvictionIterator iter;
-    uint32_t liveUntilLedger;
-
-    EvictionResultEntry(LedgerKey const& key, EvictionIterator const& iter,
-                        uint32_t liveUntilLedger)
-        : key(key), iter(iter), liveUntilLedger(liveUntilLedger)
-    {
-    }
-};
-
-struct EvictionResult
-{
-    // List of keys eligible for eviction in the order in which they occur in
-    // the bucket
-    std::list<EvictionResultEntry> eligibleKeys{};
-
-    // Eviction iterator at the end of the scan region
-    EvictionIterator endOfRegionIterator;
-
-    // LedgerSeq which this scan is based on
-    uint32_t initialLedger{};
-
-    // State archival settings that this scan is based on
-    StateArchivalSettings initialSas;
-
-    EvictionResult(StateArchivalSettings const& sas) : initialSas(sas)
-    {
-    }
-
-    // Returns true if this is a valid archival scan for the current ledger
-    // and archival settings. This is necessary because we start the scan
-    // for ledger N immediately after N - 1 closes. However, ledger N may
-    // contain a network upgrade changing eviction scan settings. Legacy SQL
-    // scans will run based on the changes that occurred during ledger N,
-    // meaning the scan we started at ledger N - 1 is invalid since it was based
-    // off of older settings.
-    bool isValid(uint32_t currLedger,
-                 StateArchivalSettings const& currSas) const;
-};
-
-struct EvictionCounters
-{
-    medida::Counter& entriesEvicted;
-    medida::Counter& bytesScannedForEviction;
-    medida::Counter& incompleteBucketScan;
-    medida::Counter& evictionCyclePeriod;
-    medida::Counter& averageEvictedEntryAge;
-
-    EvictionCounters(Application& app);
-};
-
-class EvictionStatistics
-{
-  private:
-    std::mutex mLock{};
-
-    // Only record metrics if we've seen a complete cycle to avoid noise
-    bool mCompleteCycle{false};
-    uint64_t mEvictedEntriesAgeSum{};
-    uint64_t mNumEntriesEvicted{};
-    uint32_t mEvictionCycleStartLedger{};
-
-  public:
-    // Evicted entry "age" is the delta between its liveUntilLedger and the
-    // ledger when the entry is actually evicted
-    void recordEvictedEntry(uint64_t age);
-
-    void submitMetricsAndRestartCycle(uint32_t currLedgerSeq,
-                                      EvictionCounters& counters);
-};
 
 /**
  * BucketManager is responsible for maintaining a collection of Buckets of
@@ -177,34 +58,145 @@ class EvictionStatistics
  * in temporary directories -- and then "adopted" by the BucketManager, moved
  * into its directory and managed by it.
  */
-
 class BucketManager : NonMovableOrCopyable
 {
+    template <class BucketT>
+    using BucketMapT = std::map<Hash, std::shared_ptr<BucketT>>;
+
+    template <class BucketT>
+    using FutureMapT =
+        UnorderedMap<MergeKey, std::shared_future<std::shared_ptr<BucketT>>>;
+
+    static std::string const kLockFilename;
+
+    Application& mApp;
+    std::unique_ptr<LiveBucketList> mLiveBucketList;
+    std::unique_ptr<HotArchiveBucketList> mHotArchiveBucketList;
+    std::unique_ptr<BucketSnapshotManager> mSnapshotManager;
+    std::unique_ptr<TmpDirManager> mTmpDirManager;
+    std::unique_ptr<TmpDir> mWorkDir;
+    BucketMapT<LiveBucket> mSharedLiveBuckets;
+    BucketMapT<HotArchiveBucket> mSharedHotArchiveBuckets;
+    std::shared_ptr<SearchableLiveBucketListSnapshot>
+        mSearchableBucketListSnapshot{};
+
+    // Lock for managing raw Bucket files or the bucket directory. This lock is
+    // only required for file access, but is not required for logical changes to
+    // a BucketList (i.e. addLiveBatch).
+    mutable std::recursive_mutex mBucketMutex;
+    std::unique_ptr<std::string> mLockedBucketDir;
+    medida::Meter& mBucketLiveObjectInsertBatch;
+    medida::Meter& mBucketArchiveObjectInsertBatch;
+    medida::Timer& mBucketAddLiveBatch;
+    medida::Timer& mBucketAddArchiveBatch;
+    medida::Timer& mBucketSnapMerge;
+    medida::Counter& mSharedBucketsSize;
+    medida::Meter& mBucketListDBBloomMisses;
+    medida::Meter& mBucketListDBBloomLookups;
+    medida::Counter& mLiveBucketListSizeCounter;
+    medida::Counter& mArchiveBucketListSizeCounter;
+    EvictionCounters mBucketListEvictionCounters;
+    MergeCounters mMergeCounters;
+    std::shared_ptr<EvictionStatistics> mEvictionStatistics{};
+    std::map<LedgerEntryTypeAndDurability, medida::Counter&>
+        mBucketListEntryCountCounters;
+    std::map<LedgerEntryTypeAndDurability, medida::Counter&>
+        mBucketListEntrySizeCounters;
+
+    std::future<EvictionResult> mEvictionFuture{};
+
+    bool const mDeleteEntireBucketDirInDtor;
+
+    // Records bucket-merges that are currently _live_ in some FutureBucket, in
+    // the sense of either running, or finished (with or without the
+    // FutureBucket being resolved). Entries in this map will be cleared when
+    // the FutureBucket is _cleared_ (typically when the owning BucketList level
+    // is committed).
+    FutureMapT<LiveBucket> mLiveBucketFutures;
+    FutureMapT<HotArchiveBucket> mHotArchiveBucketFutures;
+
+    // Records bucket-merges that are _finished_, i.e. have been adopted as
+    // (possibly redundant) bucket files. This is a "weak" (bi-multi-)map of
+    // hashes, that does not count towards std::shared_ptr refcounts, i.e. does
+    // not keep either the output bucket or any of its input buckets
+    // alive. Needs to be queried and updated on mSharedBuckets GC events.
+    BucketMergeMap mFinishedMerges;
+
+    std::atomic<bool> mIsShutdown{false};
+
+    void cleanupStaleFiles();
+    void deleteTmpDirAndUnlockBucketDir();
+    void deleteEntireBucketDir();
+
+    medida::Timer& recordBulkLoadMetrics(std::string const& label,
+                                         size_t numEntries) const;
+    medida::Timer& getPointLoadTimer(LedgerEntryType t) const;
+
+    void updateSharedBucketSize();
+
+    template <class BucketT>
+    std::shared_ptr<BucketT> adoptFileAsBucketInternal(
+        std::string const& filename, uint256 const& hash, MergeKey* mergeKey,
+        std::unique_ptr<BucketIndex const> index,
+        BucketMapT<BucketT>& bucketMap, FutureMapT<BucketT>& futureMap);
+
+    template <class BucketT>
+    std::shared_ptr<BucketT>
+    getBucketByHashInternal(uint256 const& hash,
+                            BucketMapT<BucketT>& bucketMap);
+    template <class BucketT>
+    std::shared_ptr<BucketT>
+    getBucketIfExistsInternal(uint256 const& hash,
+                              BucketMapT<BucketT> const& bucketMap) const;
+
+    template <class BucketT>
+    std::shared_future<std::shared_ptr<BucketT>>
+    getMergeFutureInternal(MergeKey const& key, FutureMapT<BucketT>& futureMap);
+
+    template <class BucketT>
+    void
+    putMergeFutureInternal(MergeKey const& key,
+                           std::shared_future<std::shared_ptr<BucketT>> future,
+                           FutureMapT<BucketT>& futureMap);
+
+    template <class BucketT>
+    void noteEmptyMergeOutputInternal(MergeKey const& mergeKey,
+                                      FutureMapT<BucketT>& futureMap);
+
+#ifdef BUILD_TESTS
+    bool mUseFakeTestValuesForNextClose{false};
+    uint32_t mFakeTestProtocolVersion;
+    uint256 mFakeTestBucketListHash;
+#endif
+
+  protected:
+    BucketManager(Application& app);
+    void calculateSkipValues(LedgerHeader& currentHeader);
+    std::string bucketFilename(std::string const& bucketHexHash);
+    std::string bucketFilename(Hash const& hash);
 
   public:
-    static std::unique_ptr<BucketManager> create(Application&);
+    static std::unique_ptr<BucketManager> create(Application& app);
+    virtual ~BucketManager();
 
-    virtual ~BucketManager()
-    {
-    }
-    virtual void initialize() = 0;
-    virtual void dropAll() = 0;
-    virtual std::string bucketIndexFilename(Hash const& hash) const = 0;
-    virtual std::string const& getTmpDir() = 0;
-    virtual TmpDirManager& getTmpDirManager() = 0;
-    virtual std::string const& getBucketDir() const = 0;
-    virtual LiveBucketList& getLiveBucketList() = 0;
-    virtual HotArchiveBucketList& getHotArchiveBucketList() = 0;
-    virtual BucketSnapshotManager& getBucketSnapshotManager() const = 0;
-    virtual bool renameBucketDirFile(std::filesystem::path const& src,
-                                     std::filesystem::path const& dst) = 0;
+    void initialize();
+    void dropAll();
+    std::string bucketIndexFilename(Hash const& hash) const;
+    std::string const& getTmpDir();
+    TmpDirManager& getTmpDirManager();
+    std::string const& getBucketDir() const;
+    LiveBucketList& getLiveBucketList();
+    HotArchiveBucketList& getHotArchiveBucketList();
+    BucketSnapshotManager& getBucketSnapshotManager() const;
+    bool renameBucketDirFile(std::filesystem::path const& src,
+                             std::filesystem::path const& dst);
 
-    virtual medida::Timer& getMergeTimer() = 0;
+    medida::Timer& getMergeTimer();
 
     // Reading and writing the merge counters is done in bulk, and takes a lock
     // briefly; this can be done from any thread.
-    virtual MergeCounters readMergeCounters() = 0;
-    virtual void incrMergeCounters(MergeCounters const& delta) = 0;
+    MergeCounters readMergeCounters();
+    void incrMergeCounters(MergeCounters const& delta);
 
     // Get a reference to a persistent bucket (in the BucketManager's bucket
     // directory), from the BucketManager's shared bucket-set.
@@ -217,12 +209,10 @@ class BucketManager : NonMovableOrCopyable
     // This method is mostly-threadsafe -- assuming you don't destruct the
     // BucketManager mid-call -- and is intended to be called from both main and
     // worker threads. Very carefully.
-    // Unfortunately, virtual methods cannot be templated, so we use this weird
-    // static interface to allow for a templated return type.
     template <class BucketT>
-    static std::shared_ptr<BucketT>
-    adoptFileAsBucket(BucketManager& bm, std::string const& filename,
-                      uint256 const& hash, MergeKey* mergeKey,
+    std::shared_ptr<BucketT>
+    adoptFileAsBucket(std::string const& filename, uint256 const& hash,
+                      MergeKey* mergeKey,
                       std::unique_ptr<BucketIndex const> index);
 
     // Companion method to `adoptFileAsLiveBucket` also called from the
@@ -231,137 +221,120 @@ class BucketManager : NonMovableOrCopyable
     // doesn't correspond to a file on disk; the method forgets about the
     // `FutureBucket` associated with the in-progress merge, allowing the merge
     // inputs to be GC'ed.
-    // Unfortunately, virtual methods cannot be templated, so we use this weird
-    // static interface to allow for a templated return type.
     template <class BucketT>
-    static void noteEmptyMergeOutput(BucketManager& bm,
-                                     MergeKey const& mergeKey);
+    void noteEmptyMergeOutput(MergeKey const& mergeKey);
 
     // Returns a bucket by hash if it exists and is currently managed by the
     // bucket list.
-    // Unfortunately, virtual methods cannot be templated, so we use this weird
-    // static interface to allow for a templated return type.
     template <class BucketT>
-    static std::shared_ptr<BucketT> getBucketIfExists(BucketManager const& bm,
-                                                      uint256 const& hash);
+    std::shared_ptr<BucketT> getBucketIfExists(uint256 const& hash);
 
     // Return a bucket by hash if we have it, else return nullptr.
-    // Unfortunately, virtual methods cannot be templated, so we use this weird
-    // static getter interface to allow for a templated return type.
     template <class BucketT>
-    static std::shared_ptr<BucketT> getBucketByHash(BucketManager& bm,
-                                                    uint256 const& hash);
+    std::shared_ptr<BucketT> getBucketByHash(uint256 const& hash);
 
     // Get a reference to a merge-future that's either running (or finished
     // somewhat recently) from either a map of the std::shared_futures doing the
     // merges and/or a set of records mapping merge inputs to outputs and the
     // set of outputs held in the BucketManager. Returns an invalid future if no
     // such future can be found or synthesized.
-    // Unfortunately, virtual methods cannot be templated, so we use this weird
-    // static getter interface to allow for a templated return type.
     template <class BucketT>
-    static std::shared_future<std::shared_ptr<BucketT>>
-    getMergeFuture(BucketManager& bucketManager, MergeKey const& key);
+    std::shared_future<std::shared_ptr<BucketT>>
+    getMergeFuture(MergeKey const& key);
 
     // Add a reference to a merge _in progress_ (not yet adopted as a file) to
     // the BucketManager's internal map of std::shared_futures doing merges.
     // There is no corresponding entry-removal API: the std::shared_future will
     // be removed from the map when the merge completes and the output file is
     // adopted.
-    // Unfortunately, virtual methods cannot be templated, so we use this weird
-    // static interface to allow for a templated return type.
     template <class BucketT>
-    static void
-    putMergeFuture(BucketManager& bm, MergeKey const& key,
-                   std::shared_future<std::shared_ptr<BucketT>> future);
+    void putMergeFuture(MergeKey const& key,
+                        std::shared_future<std::shared_ptr<BucketT>> future);
 
 #ifdef BUILD_TESTS
     // Drop all references to merge futures in progress.
-    virtual void clearMergeFuturesForTesting() = 0;
+    void clearMergeFuturesForTesting();
 #endif
 
     // Forget any buckets not referenced by the current BucketList. This will
     // not immediately cause the buckets to delete themselves, if someone else
     // is using them via a shared_ptr<>, but the BucketManager will no longer
     // independently keep them alive.
-    virtual void forgetUnreferencedBuckets() = 0;
+    void forgetUnreferencedBuckets();
 
     // Feed a new batch of entries to the bucket list. This interface expects to
     // be given separate init (created) and live (updated) entry vectors. The
     // `header` value should be taken from the ledger at which this batch is
     // being added.
-    virtual void addLiveBatch(Application& app, LedgerHeader header,
-                              std::vector<LedgerEntry> const& initEntries,
-                              std::vector<LedgerEntry> const& liveEntries,
-                              std::vector<LedgerKey> const& deadEntries) = 0;
-    virtual void
-    addHotArchiveBatch(Application& app, LedgerHeader header,
-                       std::vector<LedgerEntry> const& archivedEntries,
-                       std::vector<LedgerKey> const& restoredEntries,
-                       std::vector<LedgerKey> const& deletedEntries) = 0;
+    void addLiveBatch(Application& app, LedgerHeader header,
+                      std::vector<LedgerEntry> const& initEntries,
+                      std::vector<LedgerEntry> const& liveEntries,
+                      std::vector<LedgerKey> const& deadEntries);
+    void addHotArchiveBatch(Application& app, LedgerHeader header,
+                            std::vector<LedgerEntry> const& archivedEntries,
+                            std::vector<LedgerKey> const& restoredEntries,
+                            std::vector<LedgerKey> const& deletedEntries);
 
     // Update the given LedgerHeader's bucketListHash to reflect the current
     // state of the bucket list.
-    virtual void snapshotLedger(LedgerHeader& currentHeader) = 0;
+    void snapshotLedger(LedgerHeader& currentHeader);
 
     // Sets index for bucket b if b is not already indexed and if BucketManager
     // is not shutting down. In most cases, there should only be a single index
     // for each bucket. However, during startup there are race conditions where
     // a bucket may be indexed twice. If there is an index race, set index with
     // this function, otherwise use Bucket::setIndex().
-    virtual void maybeSetIndex(std::shared_ptr<Bucket> b,
-                               std::unique_ptr<BucketIndex const>&& index) = 0;
+    void maybeSetIndex(std::shared_ptr<Bucket> b,
+                       std::unique_ptr<BucketIndex const>&& index);
 
     // Scans BucketList for non-live entries to evict starting at the entry
     // pointed to by EvictionIterator. Scans until `maxEntriesToEvict` entries
     // have been evicted or maxEvictionScanSize bytes have been scanned.
-    virtual void scanForEvictionLegacy(AbstractLedgerTxn& ltx,
-                                       uint32_t ledgerSeq) = 0;
+    void scanForEvictionLegacy(AbstractLedgerTxn& ltx, uint32_t ledgerSeq);
 
-    virtual void startBackgroundEvictionScan(uint32_t ledgerSeq) = 0;
-    virtual void
-    resolveBackgroundEvictionScan(AbstractLedgerTxn& ltx, uint32_t ledgerSeq,
-                                  LedgerKeySet const& modifiedKeys) = 0;
+    void startBackgroundEvictionScan(uint32_t ledgerSeq);
+    void resolveBackgroundEvictionScan(AbstractLedgerTxn& ltx,
+                                       uint32_t ledgerSeq,
+                                       LedgerKeySet const& modifiedKeys);
 
-    virtual medida::Meter& getBloomMissMeter() const = 0;
-    virtual medida::Meter& getBloomLookupMeter() const = 0;
+    medida::Meter& getBloomMissMeter() const;
+    medida::Meter& getBloomLookupMeter() const;
 
 #ifdef BUILD_TESTS
     // Install a fake/assumed ledger version and bucket list hash to use in next
     // call to addLiveBatch and snapshotLedger. This interface exists only for
     // testing in a specific type of history replay.
-    virtual void setNextCloseVersionAndHashForTesting(uint32_t protocolVers,
-                                                      uint256 const& hash) = 0;
+    void setNextCloseVersionAndHashForTesting(uint32_t protocolVers,
+                                              uint256 const& hash);
 
     // Return the set of buckets in the current `getBucketDir()` directory.
     // This interface exists only for checking that the BucketDir isn't
     // leaking buckets, in tests.
-    virtual std::set<Hash> getBucketHashesInBucketDirForTesting() const = 0;
+    std::set<Hash> getBucketHashesInBucketDirForTesting() const;
 
-    virtual medida::Counter& getEntriesEvictedCounter() const = 0;
+    medida::Counter& getEntriesEvictedCounter() const;
 #endif
 
     // Return the set of buckets referenced by the BucketList
-    virtual std::set<Hash> getBucketListReferencedBuckets() const = 0;
+    std::set<Hash> getBucketListReferencedBuckets() const;
 
     // Return the set of buckets referenced by the BucketList, LCL HAS,
     // and publish queue.
-    virtual std::set<Hash> getAllReferencedBuckets() const = 0;
+    std::set<Hash> getAllReferencedBuckets() const;
 
     // Check for missing bucket files that would prevent `assumeState` from
     // succeeding
-    virtual std::vector<std::string>
-    checkForMissingBucketsFiles(HistoryArchiveState const& has) = 0;
+    std::vector<std::string>
+    checkForMissingBucketsFiles(HistoryArchiveState const& has);
 
     // Assume state from `has` in BucketList: find and attach all buckets in
     // `has`, set current BL.
-    virtual void assumeState(HistoryArchiveState const& has,
-                             uint32_t maxProtocolVersion,
-                             bool restartMerges) = 0;
+    void assumeState(HistoryArchiveState const& has,
+                     uint32_t maxProtocolVersion, bool restartMerges);
 
-    virtual void shutdown() = 0;
+    void shutdown();
 
-    virtual bool isShutdown() const = 0;
+    bool isShutdown() const;
 
     // Load the complete state of the ledger from the provided HAS. Throws if
     // any of the buckets referenced in the HAS do not exist.
@@ -372,13 +345,12 @@ class BucketManager : NonMovableOrCopyable
     //
     // Also note: this returns a large map -- likely multiple GB of memory on
     // public nodes. The whole ledger. Call carefully, and only offline.
-    virtual std::map<LedgerKey, LedgerEntry>
-    loadCompleteLedgerState(HistoryArchiveState const& has) = 0;
+    std::map<LedgerKey, LedgerEntry>
+    loadCompleteLedgerState(HistoryArchiveState const& has);
 
     // Merge the bucket list of the provided HAS into a single "super bucket"
     // consisting of only live entries, and return it.
-    virtual std::shared_ptr<LiveBucket>
-    mergeBuckets(HistoryArchiveState const& has) = 0;
+    std::shared_ptr<LiveBucket> mergeBuckets(HistoryArchiveState const& has);
 
     // Visits all the active ledger entries or subset thereof.
     //
@@ -397,23 +369,27 @@ class BucketManager : NonMovableOrCopyable
     // When `filterEntry` and `acceptEntry` always return `true`, this is
     // equivalent to iterating over `loadCompleteLedgerState`, so the same
     // memory/runtime implications apply.
-    virtual void visitLedgerEntries(
+    void visitLedgerEntries(
         HistoryArchiveState const& has, std::optional<int64_t> minLedger,
         std::function<bool(LedgerEntry const&)> const& filterEntry,
         std::function<bool(LedgerEntry const&)> const& acceptEntry,
-        bool includeAllStates) = 0;
+        bool includeAllStates);
 
     // Schedule a Work class that verifies the hashes of all referenced buckets
     // on background threads.
-    virtual std::shared_ptr<BasicWork>
-    scheduleVerifyReferencedBucketsWork() = 0;
+    std::shared_ptr<BasicWork> scheduleVerifyReferencedBucketsWork();
 
-    virtual Config const& getConfig() const = 0;
+    Config const& getConfig() const;
 
     // Get bucketlist snapshot
-    virtual std::shared_ptr<SearchableLiveBucketListSnapshot>
-    getSearchableLiveBucketListSnapshot() = 0;
+    std::shared_ptr<SearchableLiveBucketListSnapshot>
+    getSearchableLiveBucketListSnapshot();
 
-    virtual void reportBucketEntryCountMetrics() = 0;
+    void reportBucketEntryCountMetrics();
 };
+
+#define SKIP_1 50
+#define SKIP_2 5000
+#define SKIP_3 50000
+#define SKIP_4 500000
 }
