+++ src/bucket/BucketSnapshotManager.h
@@ -6,7 +6,7 @@
 
 #include "bucket/Bucket.h"
 #include "bucket/BucketList.h"
-#include "bucket/BucketManagerImpl.h"
+#include "bucket/BucketManager.h"
 #include "util/NonCopyable.h"
 #include "util/UnorderedMap.h"
