+++ src/bucket/test/BucketIndexTests.cpp
@@ -635,8 +635,7 @@ TEST_CASE("serialize bucket indexes", "[bucket][bucketindex]")
         auto indexFilename = test.getBM().bucketIndexFilename(bucketHash);
         REQUIRE(fs::exists(indexFilename));
 
-        auto b = BucketManager::getBucketByHash<LiveBucket>(test.getBM(),
-                                                            bucketHash);
+        auto b = test.getBM().getBucketByHash<LiveBucket>(bucketHash);
         REQUIRE(b->isIndexed());
 
         auto onDiskIndex =
@@ -662,8 +661,7 @@ TEST_CASE("serialize bucket indexes", "[bucket][bucketindex]")
         }
 
         // Check if in-memory index has correct params
-        auto b = BucketManager::getBucketByHash<LiveBucket>(test.getBM(),
-                                                            bucketHash);
+        auto b = test.getBM().getBucketByHash<LiveBucket>(bucketHash);
         REQUIRE(!b->isEmpty());
         REQUIRE(b->isIndexed());
